+++ 03/configMap/ngx-config-map.yaml
@@ -1,0 +1,32 @@
+data:
+  nginx.conf: |
+    user nginx;
+    worker_processes  1;
+    error_log  /var/log/nginx/error.log;
+    events {
+      worker_connections  10240;
+    }
+    http {
+      log_format  main
+        'remote_addr:$remote_addr\t'
+        'time_local:$time_local\t'
+        'method:$request_method\t'
+        'uri:$request_uri\t'
+        'host:$host\t'
+        'status:$status\t'
+        'bytes_sent:$body_bytes_sent\t'
+        'referer:$http_referer\t'
+        'useragent:$http_user_agent\t'
+        'forwardedfor:$http_x_forwarded_for\t'
+        'request_time:$request_time';
+      access_log  /var/log/nginx/access.log main;
+      server {
+          listen       80;
+          server_name  _;
+          location / {
+              root   /usr/share/nginx/html;
+              index  index.html index.htm;
+          }
+
+      }
+    }
