+++ 03/ngx.yaml
@@ -1,0 +1,39 @@
+# Please edit the object below. Lines beginning with a '#' will be ignored,
+# and an empty file will abort the edit. If an error occurs while saving this file will be
+# reopened with the relevant failures.
+#
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  labels:
+    run: ngx
+  name: ngx
+  namespace: default
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      run: ngx
+  template:
+    metadata:
+      labels:
+        run: ngx
+    spec:
+      containers:
+      - image: nginx
+        imagePullPolicy: Always
+        name: ngx
+---
+apiVersion: v1
+kind: Service
+metadata:
+  labels:
+    run: ngx
+  name: ngx
+spec:
+  ports:
+  - port: 80
+    protocol: TCP
+    targetPort: 80
+  selector:
+    run: ngx
