+++ 05/viz.md
@@ -1,0 +1,11 @@
+# Visualization 
+
+The basic Kubernetes GUI is pretty lame when you consider some other options that are much better.  Two to consider here are [Lens](https://k8slens.dev/) and [K9s](https://k9scli.io/). 
+
+## Lens
+
+Lens is a great native desktop application that allows you to visualize multiple Kubernetes clusters. 
+
+## K9s
+
+K9s is a TUI (text user interface) that is great for installing on bastion hosts when your primary laptop can't reach it.  Installation is simple following the instructions on [the website](https://k9scli.io/).
