+++ 06/dynamo-example/Dockerfile
@@ -1,0 +1,3 @@
+FROM tiangolo/uwsgi-nginx-flask:python3.6-alpine3.7
+RUN pip install boto3
+COPY ./app /app
