+++ 06/serverless/Dockerfile
@@ -1,0 +1,1 @@
+FROM lambci/lambda:python3.8
