+++ 06/serverless/requirements.txt
@@ -1,0 +1,1 @@
+kubernetes
