--- segment07-integrations/Dockerfile
@@ -1,1 +1,0 @@
-FROM lambci/lambda:python3.8
