--- segment07-integrations/dynamo-example/Dockerfile
@@ -1,3 +1,0 @@
-FROM tiangolo/uwsgi-nginx-flask:python3.6-alpine3.7
-RUN pip install boto3
-COPY ./app /app
