--- segment07-integrations/requirements.txt
@@ -1,1 +1,0 @@
-kubernetes
